# Conflicts:
#	src/main/java/org/oracle/s202501a/controller/sh_controller/HomeController.java
#	src/main/resources/application.yml

+++ src/main/java/org/oracle/s202501a/dto/ny_dto/Board.java
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class Board {
-    private int board_No;
+    private Long board_No;
     private Long emp_No; 
     private Timestamp modifiedDate;
     private String content;
